requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gspread==6.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.25.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-telegram-bot==21.1.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
StrEnum==0.4.15
tqdm==4.66.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
